# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: Get all financial goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalInput"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /goals/{id}/contribute:
    post:
      operationId: contributeToGoal
      tags: [goals]
      summary: Add contribution to goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalContributionInput"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
  /recurring:
    get:
      operationId: getRecurringTransactions
      tags: [recurring]
      summary: Get recurring transactions
      responses:
        "200":
          description: List of recurring transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringTransaction"
    post:
      operationId: createRecurringTransaction
      tags: [recurring]
      summary: Create recurring transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringTransactionInput"
      responses:
        "201":
          description: Created recurring transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
  /recurring/{id}:
    put:
      operationId: updateRecurringTransaction
      tags: [recurring]
      summary: Update recurring transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecurringTransactionInput"
      responses:
        "200":
          description: Updated recurring transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
    delete:
      operationId: deleteRecurringTransaction
      tags: [recurring]
      summary: Delete recurring transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cash, card, wallet, savings, investment]
        currency:
          type: string
        initialBalance:
          type: number
        currentBalance:
          type: number
        color:
          type: string
        icon:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, currency, initialBalance, currentBalance, createdAt]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, color, icon, isDefault, createdAt]
    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        color:
          type: string
        icon:
          type: string
      required: [name, type, color, icon]
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        color:
          type: string
        icon:
          type: string
        monthlySuggestion:
          type: number
        progressPercent:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, targetAmount, currentAmount, progressPercent, createdAt]
    CreateGoalInput:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        color:
          type: string
        icon:
          type: string
      required: [name, targetAmount]
    GoalContributionInput:
      type: object
      properties:
        amount:
          type: number
      required: [amount]
...
    RecurringTransaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        accountId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        dayOfMonth:
          type: integer
        nextDate:
          type: string
          format: date
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        account:
          $ref: "#/components/schemas/Account"
        category:
          $ref: "#/components/schemas/Category"
      required: [id, type, amount, accountId, categoryId, description, frequency, isActive, createdAt]
    CreateRecurringTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        accountId:
          type: integer
        categoryId:
          type: integer
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        dayOfMonth:
          type: integer
        startDate:
          type: string
          format: date
        isActive:
          type: boolean
      required: [type, amount, accountId, categoryId, description, frequency]
...