# lib/api-zod/src/generated/types/recurringTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Account } from "./account";
import type { Category } from "./category";
import type { RecurringTransactionFrequency } from "./recurringTransactionFrequency";
import type { RecurringTransactionType } from "./recurringTransactionType";

export interface RecurringTransaction {
  id: number;
  type: RecurringTransactionType;
  amount: number;
  accountId: number;
  categoryId: number;
  description: string;
  frequency: RecurringTransactionFrequency;
  dayOfMonth?: number;
  nextDate?: Date;
  isActive: boolean;
  createdAt: Date;
  account?: Account;
  category?: Category;
}